feat: 重构一键解压功能 * 大幅提高了解压速度 * 可以在解压的时候指定游戏名了（防止一些奇怪的字符串干扰识别） * BgTask化

+++ GalgameManager/Models/BgTasks/BgTaskBase.cs
@@ -25,7 +25,7 @@ public abstract class BgTaskBase
 
     public virtual bool OnSearch(string key) => false;
     
-    protected void ChangeProgress(int current, int total, string message)
+    protected void ChangeProgress(long current, long total, string message)
     {
         CurrentProgress = new Progress
         {
